+++ api/src/Application/Actions/Occasion/CreateOccasionAction.php
@@ -9,9 +9,12 @@ use App\Application\Serializable\Occasion\SerializableOccasion;
 use App\Domain\Occasion\OccasionRepository;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Log\LoggerInterface;
+use Slim\Exception\HttpBadRequestException;
 
 class CreateOccasionAction extends Action
 {
+  protected $occasionRepository;
+  
   public function __construct(
     LoggerInterface $logger,
     OccasionRepository $occasionRepository
@@ -25,6 +28,8 @@ class CreateOccasionAction extends Action
     $this->assertAuth();
     $this->assertUtilisateurAuthEstAdmin();
     $body = $this->getFormData();
+
+    if (!isset($body['titre'])) throw new HttpBadRequestException($this->request, "Le titre est obligatoire");
 
     $occasion = $this->occasionRepository->create($body['titre']);